Create 116 Maps_ Fundamental
https://replit.com/@MujtabaAbid/116-Maps-Fundamentals#index.js
Map is a data structure that we can use to map values to key.
The big difference between maps and object is keys can be of any type(Object, arrays or even maps). So in objects, keys normally in strings.
Set method is used to fill the Map
In order to read data we use get method.